requirements.txt
amqp==5.1.1
appdirs==1.4.4
asgiref==3.4.1
async-timeout==3.0.1
attrs==21.2.0
beautifulsoup4==4.11.1
billiard==3.6.4.0
celery==5.2.7
certifi==2021.5.30
charset-normalizer==2.0.6
click==8.1.3
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.2.0
constantly==15.1.0
Django==3.2.7
django-celery-beat==2.4.0
django-celery-results==2.4.0
django-channels==0.7.0
django-channels-presence==1.0.0
django-cors-headers==3.13.0
django-timezone-field==5.0
djangorestframework==3.14.0
hiredis==2.0.0
hyperlink==21.0.0
idna==3.2
incremental==21.3.0
kombu==5.2.4
lxml==4.6.3
msgpack==1.0.2
multitasking==0.0.9
numpy==1.21.2
oauthlib==3.2.2
pandas==1.5.1
prompt-toolkit==3.0.20
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.20
python-crontab==2.6.0
python-dateutil==2.8.2
pytz==2022.6
redis==3.5.3
requests==2.26.0
requests-oauthlib==1.3.1
six==1.16.0
soupsieve==2.2.1
sqlparse==0.4.2
txaio==21.2.1
typing-extensions==3.10.0.2
tzdata==2022.6
urllib3==1.26.7
vine==5.0.0
wcwidth==0.2.5
yfinance==0.1.87
zope.interface==5.4.0
